microservicio/infraestructura/src/test/java/com/ceiba/factura/controlador/ComandoControladorFacturaTest.java
package com.ceiba.factura.controlador;

import com.ceiba.ApplicationMock;
import com.ceiba.factura.comando.ComandoSolicitudFacturar;
import com.ceiba.factura.testdatabuilder.ComandoFacturaTestDataBuilder;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.http.MediaType;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import static org.hamcrest.core.Is.is;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@RunWith(SpringRunner.class)
@ContextConfiguration(classes = ApplicationMock.class)
@WebMvcTest(ComandoControladorFactura.class)
public class ComandoControladorFacturaTest {


    @Autowired
    private ObjectMapper objectMapper;

    @Autowired
    private MockMvc mocMvc;

    @Test
    void crearFactura() throws Exception {
        // arrange
        ComandoSolicitudFacturar entrada = new ComandoFacturaTestDataBuilder().getReserva().build();

        // act - assert
        var resultado =  mocMvc.perform(post("/factura/facturar")
                        .contentType(MediaType.APPLICATION_JSON)
                        .content(objectMapper.writeValueAsString(entrada)))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.valor.id", is(2)))
                .andExpect(jsonPath("$.valor.idReserva", is(2)))
                .andExpect(jsonPath("$.valor.estado", is("FINALIZADA")))
                .andReturn();
    }
}
